# packages/contract/openapi.yaml
openapi: 3.0.3
info:
  title: Cumora API
  version: 0.2.2
  description: |
    Cumora 单一契约源(ADR 0004 · 票 #48)。覆盖全部存活路由域:
    core(auth/me/companies/invitations)、computers/agents、hr(编外人事代理)、
    conversations/messages、email、projects、participants、boards、calendar、
    documents、push、uploads、projects、observability、devtools、admin、
    shipping、runtime(BYOA daemon 专用面)。

    ## WebSocket / SSE 事件面描述策略

    本规范描述 HTTP 面。长连接通道:

    - **WebSocket `/ws`** — 实时事件广播(message.new/delta、typing、participants/
      computers 状态、board/doc/calendar 变更等)。连接鉴权走 `/auth/ws-ticket`
      票据。事件载荷自 #221 起契约化:逐事件 schema 见同目录 `ws-events.json`
      (单一 JSON Schema 契约源,`npm run contract:gen` 再生 TS `src/ws-events.d.ts`
      与 Go `internal/events/ws.gen.go`,三端手写类型已退役)。
    - **SSE `/runtime/wake-stream`** — daemon 唤醒流(wake/steer 两类事件);载荷
      形状暂不入契约(#49 的 wire 级验收镜像测试承载——同一套测试跑 TS 与 Go,
      差异即回归)。

    WS 事件载荷不进本文件的 components:事件面走独立 JSON Schema 契约文件,
    避免把 $defs 级联混入 HTTP 生成管线(oapi-codegen / openapi-typescript 输出
    保持面干净);两文件通过 `openapi.yaml#/components/schemas/*` $ref 共享组件。

    ## 鉴权

    - `sessionBearer` — 用户会话(Authorization: Bearer <token> + X-Company-Id 租户头)
    - `agentRuntimeJWT` — BYOA daemon 的 per-agent runtime JWT(/runtime/* 全域)
    - `deviceToken` — daemon 设备凭证(/api/computers/pair 等)
    - 部分路由匿名(public/livez/og/auth 回调)
servers:
  - url: /
  - url: http://localhost:5181
...
security:
  - sessionBearer: []
paths:
...
  /api/computers:
    get:
      tags: [computers]
      operationId: getComputers
      summary: 本公司计算机(遗留 cloud 行已由 kind 守卫排除)
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Computer' }
    post:
      tags: [computers]
      operationId: requestPairingCode
      summary: 申请配对码(持久公司级)
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '201':
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [code]
                properties:
                  code: { type: string }
                  expiresInSeconds: { type: integer, nullable: true }
  /api/computers/{id}:
    delete:
      tags: [computers]
      operationId: deleteComputer
      summary: 吊销计算机(设备令牌+派生 JWT 全失效)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
  /api/computers/{id}/repair:
    post:
      tags: [computers]
      operationId: repairComputer
      summary: 重连码(保留计算机与其 agent)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [code]
                properties:
                  code: { type: string }
                  expiresInSeconds: { type: integer, nullable: true }
  /api/computers/pair:
    post:
      tags: [computers]
      operationId: pairComputer
      security: [{ deviceToken: [] }]
      summary: daemon 兑换配对码 → 设备凭证
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
                hostName: { type: string }
                engines: { type: array, items: { type: string } }
                version: { type: string }
                supervised: { type: boolean }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [computerId, deviceToken]
                properties:
                  computerId: { type: string }
                  deviceToken: { type: string }
  /api/computers/me/agents:
    get:
      tags: [computers]
      operationId: listAgentsForComputer
      security: [{ deviceToken: [] }]
      summary: daemon 拉取其托管的 agent 名册
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { type: array, items: { type: object, additionalProperties: true } }
  /api/computers/heartbeat:
    post:
      tags: [computers]
      operationId: heartbeatComputer
      security: [{ deviceToken: [] }]
      summary: daemon 心跳(30s)
      requestBody:
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
  /api/computers/me/skills:
    get:
      tags: [computers]
      operationId: listComputerSkills
      security: [{ deviceToken: [] }]
      summary: '#261 daemon 拉取其托管 agent 所属公司的 skills 清单(带 bundle_hash 增量键)'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [companyId, name, description, bundleHash]
                  properties:
                    companyId: { type: string }
                    name: { type: string }
                    description: { type: string }
                    bundleHash: { type: string }
  /api/computers/me/project-report:
    post:
      tags: [computers]
      operationId: reportProjectChanges
      security: [{ deviceToken: [] }]
      summary: '#337 daemon watcher 上报挂载工作区文件变更(去抖批量;server 对账已知态→快照→广播)'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  maxItems: 500
                  items:
                    type: object
                    required: [workspaceId, path]
                    properties:
                      projectId: { type: string }
                      path: { type: string }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  changed: { type: integer }
  /api/computers/me/skills/{hash}:
    get:
      tags: [computers]
      operationId: getComputerSkillBundle
      security: [{ deviceToken: [] }]
      summary: '#261 daemon 按内容哈希拉取整包技能文件集'
      parameters:
        - { name: hash, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [name, files]
                properties:
                  name: { type: string }
                  files:
                    type: array
                    items:
                      type: object
                      required: [path, body]
                      properties:
                        path: { type: string }
                        body: { type: string }
  /api/agents/{id}/runtime-token:
    post:
      tags: [agents]
      operationId: mintAgentRuntimeToken
      security: [{ deviceToken: [] }]
      summary: daemon 为 agent 铸 2h runtime JWT
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [token, expiresAt]
                properties:
                  token: { type: string }
                  expiresAt: { type: string, format: date-time }
...
  /api/agents/{id}/computer:
    post:
      tags: [agents]
      operationId: assignAgentComputer
      summary: 指派计算机+引擎
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [computerId]
              properties:
                computerId: { type: string }
                engine: { $ref: '#/components/schemas/EngineId' }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [ok, kind, engine]
                properties:
                  ok: { type: boolean }
                  kind: { $ref: '#/components/schemas/ComputerKind' }
                  engine: { $ref: '#/components/schemas/EngineId' }
...
components:
  securitySchemes:
    sessionBearer:
      type: http
      scheme: bearer
      description: 用户会话令牌;多租户路由另需 X-Company-Id 头
    agentRuntimeJWT:
      type: http
      scheme: bearer
      description: daemon 经 /api/agents/:id/runtime-token 领取的 per-agent JWT(2h)
    deviceToken:
      type: http
      scheme: bearer
      description: daemon 配对所得设备凭证(computers.pair 发放)
  schemas:
...
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    ComputerKind: { type: string, enum: [local, vps] }
    ComputerStatus: { type: string, enum: [online, offline, busy] }
    EngineId: { type: string, enum: [claude, codex, grok, cursor, zcode] }
...
    Computer:
      type: object
      required: [id, company_id, owner_user_id, name, kind, available_engines, status]
      properties:
        id: { type: string }
        company_id: { type: string }
        owner_user_id: { type: string, nullable: true }
        name: { type: string }
        kind: { $ref: '#/components/schemas/ComputerKind' }
        available_engines:
          type: array
          items: { $ref: '#/components/schemas/EngineId' }
        status: { $ref: '#/components/schemas/ComputerStatus' }
        last_seen_at: { type: string, nullable: true, format: date-time }
        paired_at: { type: string, nullable: true, format: date-time }
        daemon_version: { type: string, nullable: true }
        daemon_supervised: { type: boolean, nullable: true }
        latest_daemon_version: { type: string, nullable: true }
        daemon_outdated: { type: boolean }
    # ── 以下人读域为 handler 手工映射,camelCase(与现役 web 类型一致)──
...